requirements.txt
mysqlclient==1.4.6
Django==3.0
djangorestframework==3.11.0
requests==2.22.0

